+++ src/realm/util/scope_exit.hpp
@@ -19,19 +19,27 @@
 #ifndef REALM_UTIL_SCOPE_EXIT_HPP
 #define REALM_UTIL_SCOPE_EXIT_HPP
 
+#include <exception>
 #include <optional>
 
-namespace realm {
-namespace util {
+namespace realm::util {
 
+// A guard which invokes the given function when exiting the scope (either via
+// an exception or normal flow), used to clean up state which is not owned by
+// an explicit RAII type.
+//
+// void foo()
+// {
+//     begin_foo();
+//     ScopeExit cleanup([&]() noexcept {
+//         end_foo();
+//     });
+//
+//     // Do some things which may throw an exception
+// }
 template <class H>
 class ScopeExit {
 public:
-    explicit ScopeExit(const H& handler) noexcept(std::is_nothrow_copy_constructible<H>::value)
-        : m_handler(handler)
-    {
-    }
-
     explicit ScopeExit(H&& handler) noexcept(std::is_nothrow_move_constructible<H>::value)
         : m_handler(std::move(handler))
     {
@@ -61,13 +69,39 @@ private:
 };
 
 template <class H>
-ScopeExit<typename std::remove_reference<H>::type> make_scope_exit(H&& handler) noexcept(
-    noexcept(ScopeExit<typename std::remove_reference<H>::type>(std::forward<H>(handler))))
+ScopeExit(H&&) -> ScopeExit<typename std::remove_reference_t<H>>;
+
+// Similar to ScopeExit, but the handler is *only* invoked if the scope is
+// exited via throwing an exception.
+template <class H>
+class ScopeExitFail : public ScopeExit<H> {
+public:
+    explicit ScopeExitFail(H&& handler) noexcept(std::is_nothrow_move_constructible<H>::value)
+        : ScopeExit<H>(std::move(handler))
+    {
+    }
+
+    ~ScopeExitFail()
+    {
+        if (std::uncaught_exceptions() == m_exception_count)
+            this->cancel();
+    }
+
+private:
+    int m_exception_count = std::uncaught_exceptions();
+};
+
+template <class H>
+ScopeExitFail(H&&) -> ScopeExitFail<typename std::remove_reference_t<H>>;
+
+// A helper which was required pre-C++17. New code should prefer `ScopeExit cleanup([&]() noexcept { ... })`.
+template <class H>
+ScopeExit<typename std::remove_reference_t<H>> make_scope_exit(H&& handler) noexcept(
+    noexcept(ScopeExit<typename std::remove_reference_t<H>>(std::forward<H>(handler))))
 {
-    return ScopeExit<typename std::remove_reference<H>::type>(std::forward<H>(handler));
+    return ScopeExit<typename std::remove_reference_t<H>>(std::forward<H>(handler));
 }
 
-} // namespace util
-} // namespace realm
+} // namespace realm::util
 
 #endif // REALM_UTIL_SCOPE_EXIT_HPP